Add files via upload
A batch version of the Clock App on your phone. Does not have all features. Don't know why I made it but I was bored. I'm not proud of it as in the end it is not contained into one file as it creates a temporary one in the end. Improvements on that are very welcome!